==> 0.tweetExtraction_pythonCode.ipynb <==
pip install GetOldTweets3

#sample query
#https://pypi.org/project/GetOldTweets3/
#GetOldTweets3 --querysearch ' “Nexplanon” ' --since 2006-01-01 --until 2019-12-01 --lan en --output test.csv --maxtweets 10


#Category: IUD
GetOldTweets3 --querysearch ' “Intrauterine device” ' --since 2006-01-01 --until 2006-12-31 --lan en --output IntrauterineDevice2006.csv
GetOldTweets3 --querysearch ' “Intrauterine device” ' --since 2007-01-01 --until 2007-12-31 --lan en --output IntrauterineDevice2007.csv
GetOldTweets3 --querysearch ' “Intrauterine device” ' --since 2008-01-01 --until 2008-12-31 --lan en --output IntrauterineDevice2008.csv
GetOldTweets3 --querysearch ' “Intrauterine device” ' --since 2009-01-01 --until 2009-12-31 --lan en --output IntrauterineDevice2009.csv
GetOldTweets3 --querysearch ' “Intrauterine device” ' --since 2010-01-01 --until 2010-12-31 --lan en --output IntrauterineDevice2010.csv
GetOldTweets3 --querysearch ' “Intrauterine device” ' --since 2011-01-01 --until 2011-12-31 --lan en --output IntrauterineDevice2011.csv
GetOldTweets3 --querysearch ' “Intrauterine device” ' --since 2012-01-01 --until 2012-12-31 --lan en --output IntrauterineDevice2012.csv
GetOldTweets3 --querysearch ' “Intrauterine device” ' --since 2013-01-01 --until 2013-12-31 --lan en --output IntrauterineDevice2013.csv
GetOldTweets3 --querysearch ' “Intrauterine device” ' --since 2014-01-01 --until 2014-12-31 --lan en --output IntrauterineDevice2014.csv
GetOldTweets3 --querysearch ' “Intrauterine device” ' --since 2015-01-01 --until 2015-12-31 --lan en --output IntrauterineDevice2015.csv
GetOldTweets3 --querysearch ' “Intrauterine device” ' --since 2016-01-01 --until 2016-12-31 --lan en --output IntrauterineDevice2016.csv
GetOldTweets3 --querysearch ' “Intrauterine device” ' --since 2017-01-01 --until 2017-12-31 --lan en --output IntrauterineDevice2017.csv
GetOldTweets3 --querysearch ' “Intrauterine device” ' --since 2018-01-01 --until 2018-12-31 --lan en --output IntrauterineDevice2018.csv
GetOldTweets3 --querysearch ' “Intrauterine device” ' --since 2019-01-01 --until 2019-12-01 --lan en --output IntrauterineDevice2019.csv

GetOldTweets3 --querysearch ' “IUD” ' --since 2006-01-01 --until 2006-12-31 --lan en --output IUD2006.csv
GetOldTweets3 --querysearch ' “IUD” ' --since 2007-01-01 --until 2007-12-31 --lan en --output IUD2007.csv
GetOldTweets3 --querysearch ' “IUD” ' --since 2008-01-01 --until 2008-12-31 --lan en --output IUD2008.csv
GetOldTweets3 --querysearch ' “IUD” ' --since 2009-01-01 --until 2009-12-31 --lan en --output IUD2009.csv
GetOldTweets3 --querysearch ' “IUD” ' --since 2010-01-01 --until 2010-12-31 --lan en --output IUD2010.csv
GetOldTweets3 --querysearch ' “IUD” ' --since 2011-01-01 --until 2011-12-31 --lan en --output IUD2011.csv
GetOldTweets3 --querysearch ' “IUD” ' --since 2012-01-01 --until 2012-12-31 --lan en --output IUD2012.csv
GetOldTweets3 --querysearch ' “IUD” ' --since 2013-01-01 --until 2013-12-31 --lan en --output IUD2013.csv
GetOldTweets3 --querysearch ' “IUD” ' --since 2014-01-01 --until 2014-12-31 --lan en --output IUD2014.csv
GetOldTweets3 --querysearch ' “IUD” ' --since 2015-01-01 --until 2015-12-31 --lan en --output IUD2015.csv
GetOldTweets3 --querysearch ' “IUD” ' --since 2016-01-01 --until 2016-12-31 --lan en --output IUD2016.csv
GetOldTweets3 --querysearch ' “IUD” ' --since 2017-01-01 --until 2017-12-31 --lan en --output IUD2017.csv
GetOldTweets3 --querysearch ' “IUD” ' --since 2018-01-01 --until 2018-12-31 --lan en --output IUD2018.csv
GetOldTweets3 --querysearch ' “IUD” ' --since 2019-01-01 --until 2019-12-01 --lan en --output IUD2019.csv
GetOldTweets3 --querysearch ' “Intrauterine system” ' --since 2006-01-01 --until 2019-12-01 --lan en --output IntrauterineSystem.csv
GetOldTweets3 --querysearch ' “uterine jewelry” ' --since 2006-01-01 --until 2019-12-01 --lan en --output UterineJewelry.csv

#Category: Copper IUD
GetOldTweets3 --querysearch ' “Copper-T” ' --since 2006-01-01 --until 2019-12-01 --lan en --output Copper-T.csv
GetOldTweets3 --querysearch ' “Cu-IUD” ' --since 2006-01-01 --until 2019-12-01 --lan en --output Cu-IUD.csv
GetOldTweets3 --querysearch ' “Paragard” ' --since 2006-01-01 --until 2019-12-01 --lan en --output Paragard.csv
GetOldTweets3 --querysearch ' “Paraguard” ' --since 2006-01-01 --until 2019-12-01 --lan en --output Paraguard.csv
GetOldTweets3 --querysearch ' “Copper IUD” ' --since 2006-01-01 --until 2019-12-01 --lan en --output CopperIUD.csv
GetOldTweets3 --querysearch ' “Copper IUS” ' --since 2006-01-01 --until 2019-12-01 --lan en --output CopperIUS.csv
GetOldTweets3 --querysearch ' “Copper Intrauterine device” ' --since 2006-01-01 --until 2019-12-01 --lan en --output CopperIntrauterineDevice.csv
GetOldTweets3 --querysearch ' “Copper T birth control” ' --since 2006-01-01 --until 2019-12-01 --lan en --output CopperTBirthControl.csv
GetOldTweets3 --querysearch ' “Copper T contraceptive” ' --since 2006-01-01 --until 2019-12-01 --lan en --output CopperTContraceptive.csv

#Category: LNG-IUD
GetOldTweets3 --querysearch ' “Hormonal IUD” ' --since 2006-01-01 --until 2019-12-01 --lan en --output HormonalIUD.csv
GetOldTweets3 --querysearch ' “Hormone IUD” ' --since 2006-01-01 --until 2019-12-01 --lan en --output HormoneIUD.csv
GetOldTweets3 --querysearch ' “Hormonal IUS” ' --since 2006-01-01 --until 2019-12-01 --lan en --output HormonalIUS.csv
GetOldTweets3 --querysearch ' “Hormone IUS” ' --since 2006-01-01 --until 2019-12-01 --lan en --output HormoneIUS.csv
GetOldTweets3 --querysearch ' “Hormonal Intrauterine device” ' --since 2006-01-01 --until 2019-12-01 --lan en --output HormonalIntrauterineDevice.csv
GetOldTweets3 --querysearch ' “Hormone Intrauterine device” ' --since 2006-01-01 --until 2019-12-01 --lan en --output HormoneIntrauterineDevice.csv
GetOldTweets3 --querysearch ' “Hormonal Intrauterine system” ' --since 2006-01-01 --until 2019-12-01 --lan en --output HormonalIntrauterineSystem.csv
GetOldTweets3 --querysearch ' “Hormone Intrauterine system” ' --since 2006-01-01 --until 2019-12-01 --lan en --output HormoneIntrauterineSystem.csv
GetOldTweets3 --querysearch ' “Progestin IUD” ' --since 2006-01-01 --until 2019-12-01 --lan en --output ProgestinIUD.csv
GetOldTweets3 --querysearch ' “Progesterone IUD” ' --since 2006-01-01 --until 2019-12-01 --lan en --output ProgesteroneIUD.csv
GetOldTweets3 --querysearch ' “Progesterone IUS” ' --since 2006-01-01 --until 2019-12-01 --lan en --output ProgesteroneIUS.csv
GetOldTweets3 --querysearch ' “Levonorgestrel IUD” ' --since 2006-01-01 --until 2019-12-01 --lan en --output LevonorgestrelIUD.csv
GetOldTweets3 --querysearch ' “Levonorgestrel IUS” ' --since 2006-01-01 --until 2019-12-01 --lan en --output LevonorgestrelIUS.csv
GetOldTweets3 --querysearch ' “Levonorgestrel Intrauterine device” ' --since 2006-01-01 --until 2019-12-01 --lan en --output LevonorgestrelIntrauterineDevice.csv
GetOldTweets3 --querysearch ' “Levonorgestrel Intrauterine system” ' --since 2006-01-01 --until 2019-12-01 --lan en --output LevonorgestrelIntrauterineSystem.csv
GetOldTweets3 --querysearch ' “LNG-IUD” ' --since 2006-01-01 --until 2019-12-01 --lan en --output LNG-IUD.csv
GetOldTweets3 --querysearch ' “LNG-IUS” ' --since 2006-01-01 --until 2019-12-01 --lan en --output LNG-IUS.csv
GetOldTweets3 --querysearch ' “LNG IUD” ' --since 2006-01-01 --until 2019-12-01 --lan en --output LNGIUD.csv
GetOldTweets3 --querysearch ' “LNG IUS” ' --since 2006-01-01 --until 2019-12-01 --lan en --output LNGIUS.csv
GetOldTweets3 --querysearch ' “Mirena IUD” ' --since 2006-01-01 --until 2019-12-01 --lan en --output MirenaIUD.csv
GetOldTweets3 --querysearch ' “Mirena IUS” ' --since 2006-01-01 --until 2019-12-01 --lan en --output MirenaIUS.csv
GetOldTweets3 --querysearch ' “Mirena Intrauterine device” ' --since 2006-01-01 --until 2019-12-01 --lan en --output MirenaIntrauterineDevice.csv
GetOldTweets3 --querysearch ' “Mirena Intrauterine system” ' --since 2006-01-01 --until 2019-12-01 --lan en --output MirenaIntrauterineSystem.csv
GetOldTweets3 --querysearch ' “Skyla IUD” ' --since 2006-01-01 --until 2019-12-01 --lan en --output SkylaIUD.csv
GetOldTweets3 --querysearch ' “Skyla IUS” ' --since 2006-01-01 --until 2019-12-01 --lan en --output SkylaIUS.csv
GetOldTweets3 --querysearch ' “Skyla Intrauterine device” ' --since 2006-01-01 --until 2019-12-01 --lan en --output SkylaIntrauterineDevice.csv
GetOldTweets3 --querysearch ' “Skyla Intrauterine system” ' --since 2006-01-01 --until 2019-12-01 --lan en --output SkylaIntrauterineSystem.csv
GetOldTweets3 --querysearch ' “Kyleena IUD” ' --since 2006-01-01 --until 2019-12-01 --lan en --output KyleenaIUD.csv
GetOldTweets3 --querysearch ' “Kyleena IUS” ' --since 2006-01-01 --until 2019-12-01 --lan en --output KyleenaIUS.csv
GetOldTweets3 --querysearch ' “Kyleena Intrauterine device” ' --since 2006-01-01 --until 2019-12-01 --lan en --output KyleenaIntrauterineDevice.csv
GetOldTweets3 --querysearch ' “Kyleena Intrauterine system” ' --since 2006-01-01 --until 2019-12-01 --lan en --output KyleenaIntrauterineSystem.csv
GetOldTweets3 --querysearch ' “Liletta IUD” ' --since 2006-01-01 --until 2019-12-01 --lan en --output LilettaIUD.csv
GetOldTweets3 --querysearch ' “Liletta IUS” ' --since 2006-01-01 --until 2019-12-01 --lan en --output LilettaIUS.csv
GetOldTweets3 --querysearch ' “Liletta Intrauterine device” ' --since 2006-01-01 --until 2019-12-01 --lan en --output LilettaIntrauterineDevice.csv
GetOldTweets3 --querysearch ' “Liletta Intrauterine system” ' --since 2006-01-01 --until 2019-12-01 --lan en --output LilettaIntrauterineSystem.csv

#Category: Implant
GetOldTweets3 --querysearch ' “Nexplanon” ' --since 2006-01-01 --until 2019-12-01 --lan en --output Nexplanon.csv
GetOldTweets3 --querysearch ' “Implanon” ' --since 2006-01-01 --until 2019-12-01 --lan en --output Implanon.csv
GetOldTweets3 --querysearch ' “Arm Implant” ' --since 2006-01-01 --until 2019-12-01 --lan en --output ArmImplant.csv
GetOldTweets3 --querysearch ' “BC Implant” ' --since 2006-01-01 --until 2019-12-01 --lan en --output BCImplant.csv
GetOldTweets3 --querysearch ' “B.C. Implant” ' --since 2006-01-01 --until 2019-12-01 --lan en --output B_C_Implant.csv
GetOldTweets3 --querysearch ' “Birth control Implant” ' --since 2006-01-01 --until 2019-12-01 --lan en --output BirthControlImplant.csv
GetOldTweets3 --querysearch ' “Contraception Implant” ' --since 2006-01-01 --until 2019-12-01 --lan en --output ContraceptionImplant.csv
GetOldTweets3 --querysearch ' “Contraceptive Implant” ' --since 2006-01-01 --until 2019-12-01 --lan en --output ContraceptiveImplant.csv
GetOldTweets3 --querysearch ' “Birth control Rod” ' --since 2006-01-01 --until 2019-12-01 --lan en --output BirthControlRod.csv
GetOldTweets3 --querysearch ' “Contraception Rod” ' --since 2006-01-01 --until 2019-12-01 --lan en --output ContraceptionRod.csv
GetOldTweets3 --querysearch ' “Contraceptive Rod” ' --since 2006-01-01 --until 2019-12-01 --lan en --output ContraceptiveRod.csv

#Category: The Pill 
GetOldTweets3 --querysearch ' “OC pill” ' --since 2006-01-01 --until 2019-12-01 --lan en --output OCPill.csv
GetOldTweets3 --querysearch ' “O.C. pill” ' --since 2006-01-01 --until 2019-12-01 --lan en --output O_C_Pill.csv
GetOldTweets3 --querysearch ' “BC pill” ' --since 2006-01-01 --until 2019-12-01 --lan en --output BCPill.csv
GetOldTweets3 --querysearch ' “B.C. pill” ' --since 2006-01-01 --until 2019-12-01 --lan en --output B_C_Pill.csv
GetOldTweets3 --querysearch ' “CHC pill” ' --since 2006-01-01 --until 2019-12-01 --lan en --output CHCPill.csv
GetOldTweets3 --querysearch ' “COC pill” ' --since 2006-01-01 --until 2019-12-01 --lan en --output COCPill.csv
GetOldTweets3 --querysearch ' “Birth control pill” ' --since 2006-01-01 --until 2019-12-01 --lan en --output BirthControlPill.csv
GetOldTweets3 --querysearch ' “Contraceptive pill” ' --since 2006-01-01 --until 2019-12-01 --lan en --output ContraceptivePill.csv
GetOldTweets3 --querysearch ' “Oral Contraceptive” ' --since 2006-01-01 --until 2019-12-01 --lan en --output OralContraceptive.csv
GetOldTweets3 --querysearch ' “Oral Contraception” ' --since 2006-01-01 --until 2019-12-01 --lan en --output OralContraception.csv
GetOldTweets3 --querysearch ' “Oral birth control” ' --since 2006-01-01 --until 2019-12-01 --lan en --output OralBirthControl.csv
GetOldTweets3 --querysearch ' “Combined hormonal contraceptive” ' --since 2006-01-01 --until 2019-12-01 --lan en --output CombinedHormonalContraceptive.csv
GetOldTweets3 --querysearch ' “Combined BCP” ' --since 2006-01-01 --until 2019-12-01 --lan en --output CombinedBCP.csv
GetOldTweets3 --querysearch ' “Combined OCP” ' --since 2006-01-01 --until 2019-12-01 --lan en --output CombinedOCP.csv
GetOldTweets3 --querysearch ' “Combined B.C.P.” ' --since 2006-01-01 --until 2019-12-01 --lan en --output CombinedB_C_P.csv
GetOldTweets3 --querysearch ' “Progestin only pill” ' --since 2006-01-01 --until 2019-12-01 --lan en --output ProgestinOnlyPill.csv
GetOldTweets3 --querysearch ' “Progestin BCP” ' --since 2006-01-01 --until 2019-12-01 --lan en --output ProgestinBCP.csv
GetOldTweets3 --querysearch ' “Progestin OCP” ' --since 2006-01-01 --until 2019-12-01 --lan en --output ProgestinOCP.csv

#Category: The Patch 
GetOldTweets3 --querysearch ' “BC patch” ' --since 2006-01-01 --until 2019-12-01 --lan en --output BCPatch.csv
GetOldTweets3 --querysearch ' “Birth control patch” ' --since 2006-01-01 --until 2019-12-01 --lan en --output BirthControlPatch.csv
GetOldTweets3 --querysearch ' “Contraceptive patch” ' --since 2006-01-01 --until 2019-12-01 --lan en --output ContraceptivePatch.csv
GetOldTweets3 --querysearch ' “Contraception patch” ' --since 2006-01-01 --until 2019-12-01 --lan en --output ContraceptionPatch.csv
GetOldTweets3 --querysearch ' “Ortho evra” ' --since 2006-01-01 --until 2019-12-01 --lan en --output OrthoEvra.csv
GetOldTweets3 --querysearch ' “Xulane” ' --since 2006-01-01 --until 2019-12-01 --lan en --output Xulane.csv

#Category: The Ring
GetOldTweets3 --querysearch ' “Nuva ring” ' --since 2006-01-01 --until 2019-12-01 --lan en --output Nuva_Ring.csv
GetOldTweets3 --querysearch ' “Nuvaring” ' --since 2006-01-01 --until 2019-12-01 --lan en --output Nuvaring.csv
GetOldTweets3 --querysearch ' “Nuva-ring” ' --since 2006-01-01 --until 2019-12-01 --lan en --output Nuva-ring.csv
GetOldTweets3 --querysearch ' “Contraceptive ring” ' --since 2006-01-01 --until 2019-12-01 --lan en --output ContraceptiveRing.csv
GetOldTweets3 --querysearch ' “Contraception ring” ' --since 2006-01-01 --until 2019-12-01 --lan en --output ContraceptionRing.csv
GetOldTweets3 --querysearch ' “Birth control ring” ' --since 2006-01-01 --until 2019-12-01 --lan en --output BirthControlRing.csv
GetOldTweets3 --querysearch ' “Vaginal ring” ' --since 2006-01-01 --until 2019-12-01 --lan en --output VaginalRing.csv
GetOldTweets3 --querysearch ' “Annovera” ' --since 2006-01-01 --until 2019-12-01 --lan en --output Annovera.csv

#Category: The Shot
GetOldTweets3 --querysearch ' “Birth control Shot” ' --since 2006-01-01 --until 2019-12-01 --lan en --output BirthControlShot.csv
GetOldTweets3 --querysearch ' “Contraceptive Shot” ' --since 2006-01-01 --until 2019-12-01 --lan en --output ContraceptiveShot.csv
GetOldTweets3 --querysearch ' “Contraception Shot” ' --since 2006-01-01 --until 2019-12-01 --lan en --output ContraceptionShot.csv
GetOldTweets3 --querysearch ' “Birth control injection” ' --since 2006-01-01 --until 2019-12-01 --lan en --output BirthControlInjection.csv
GetOldTweets3 --querysearch ' “Contraceptive injection” ' --since 2006-01-01 --until 2019-12-01 --lan en --output ContraceptiveInjection.csv
GetOldTweets3 --querysearch ' “Contraception Injection” ' --since 2006-01-01 --until 2019-12-01 --lan en --output ContraceptionInjection.csv
GetOldTweets3 --querysearch ' “depo provera” ' --since 2006-01-01 --until 2019-12-01 --lan en --output DepoProvera.csv
GetOldTweets3 --querysearch ' “depo-provera” ' --since 2006-01-01 --until 2019-12-01 --lan en --output Depo-Provera.csv
GetOldTweets3 --querysearch ' “depo shot” ' --since 2006-01-01 --until 2019-12-01 --lan en --output DepoShot.csv
GetOldTweets3 --querysearch ' “depo injection” ' --since 2006-01-01 --until 2019-12-01 --lan en --output DepoInjection.csv
GetOldTweets3 --querysearch ' “depo BC” ' --since 2006-01-01 --until 2019-12-01 --lan en --output DepoBC.csv
GetOldTweets3 --querysearch ' “depo birth control” ' --since 2006-01-01 --until 2019-12-01 --lan en --output DepoBirthControl.csv
GetOldTweets3 --querysearch ' “depo contraception” ' --since 2006-01-01 --until 2019-12-01 --lan en --output DepoContraception.csv
GetOldTweets3 --querysearch ' “depo contraceptive” ' --since 2006-01-01 --until 2019-12-01 --lan en --output DepoContraceptive.csv
GetOldTweets3 --querysearch ' “DMPA” ' --since 2006-01-01 --until 2019-12-01 --lan en --output DMPA.csv
GetOldTweets3 --querysearch ' “medroxyprogesterone acetate” ' --since 2006-01-01 --until 2019-12-01 --lan en --output MedroxyprogesteroneAcetate.csv
GetOldTweets3 --querysearch ' ”depot provera” ' --since 2006-01-01 --until 2019-12-01 --lan en --output DepotProvera.csv
GetOldTweets3 --querysearch ' “depot shot” ' --since 2006-01-01 --until 2019-12-01 --lan en --output DepotShot.csv
GetOldTweets3 --querysearch ' “depot injection” ' --since 2006-01-01 --until 2019-12-01 --lan en --output DepotInjection.csv
GetOldTweets3 --querysearch ' “depot birth control” ' --since 2006-01-01 --until 2019-12-01 --lan en --output DepotBirthControl.csv
GetOldTweets3 --querysearch ' “depot contraception” ' --since 2006-01-01 --until 2019-12-01 --lan en --output DepotContraception.csv
GetOldTweets3 --querysearch ' “depot contraceptive” ' --since 2006-01-01 --until 2019-12-01 --lan en --output DepotContraceptive.csv
