==> 1.tweetCleaning.ipynb <==
# Reading the tweets extracted and filtering

This notebook contains the different functions developed to:
- unify the files of the same category but extracted based on a different keyword
- remove unnecesary characters, get rid of URLS, hastaghs, etc. 
- remove tweets where the username is the keyword
- remove tweets with keywords from multiple cateogories

When running this notebook, we saved the output at two different step, first, before removing those tweets that contain keywords from multiple categories (semiclean) and the final ones after all the filtering steps (cleanTweets). 

#####################
# Load Libraries    #
#####################
#install.packages(c("tidyverse","dplyr"))
library("dplyr")
library("tidyverse")
library("ggplot2")
rm(list=ls())

#####################
# Unify Files       #
#####################
unifyFiles <- function( folder_path, folder_name ){
  
  # create a unique file with all the tweets present in that folder
  files <- list.files( paste0( folder_path, folder_name ))
  
  for( i in 1:length( files ) ){
    
    print( paste0( "Reading the file ", files[i] ) ) 
    if( i == 1 ){
      tweets <- read.csv(paste0(folder_path, folder_name, files[i]), sep = ",")
      print( paste0( "For the keyword ", files[i], " there are a total of ", nrow( tweets)))
    }else{
      new_tweet <- read.csv( paste0(folder_path, folder_name, files[i]), sep = ",")
      print( paste0( "For the keyword ", files[i], " there are a total of ", nrow( new_tweet)))
      tweets<- rbind( tweets , new_tweet )
    }
    
  }
  return( tweets)
}

##################################################################
# Clean the text in the tweets to prepare for sentiment analysis #
##################################################################
cleanTweets <- function( tweets.df ){
    tweets.df$text <- gsub("&amp;", "and", tweets.df$text) # Replace & symbol with "and" when coded as &amp;
    tweets.df$text <- gsub("&gt;", "", tweets.df$text ) # Get rid of  > symbol when coded as &gt;
    tweets.df$text <- gsub("&lt;", "", tweets.df$text ) # Get rid of  < symbol when coded as &lt;
    tweets.df$text <- gsub("&#13;", " ", tweets.df$text ) # Get rid of  html symbol for Carriage Return &#13;
    tweets.df$text <- str_replace_all(tweets.df$text ," "," ")          #Get rid of unnecessary spaces
    tweets.df$text <- gsub("“", "", tweets.df$text) # Get rid of quotes;
    tweets.df$text <- gsub("”", "", tweets.df$text) # Get rid of quotes;
    tweets.df$text <- gsub("\"", "", tweets.df$text) # Get rid of quotes;
    tweets.df$text <- gsub("—", "", tweets.df$text) # Get rid of quotes;
    tweets.df$text <- str_replace_all(tweets.df$text ,"#[a-z,A-Z]*","") # Get rid of hashtags
    tweets.df$text <- gsub("(RT|via)((?:\\b\\W*@\\w+)+)", "",tweets.df$text )# Take out retweet header / via @
    tweets.df$text <- str_replace_all(tweets.df$text,"@[a-z,A-Z]*","")  # Get rid of references to other screennames
    tweets.df$text <- gsub("http[[:alnum:][:punct:]]*", "", tweets.df$text ) # Get rid of URLs
    tweets.df$text <- gsub("www[[:alnum:][:punct:]]*", "", tweets.df$text ) # Get rid of URLs that do not start with http
    tweets.df$text <- gsub("pic.twitter[[:alnum:][:punct:]]*", "", tweets.df$text ) # Get rid of pic.twitter
    tweets.df$text <- trimws( tweets.df$text ) # Get rid of leading and/or trailing whitespace 
    return( tweets.df)
}

######################################################
# To remove tweets where the username is the keyword #
######################################################
usernameClean <- function( tweets.df, category ){
  
  tweets.df$text <- tolower( tweets.df$text )
  tweets.df$username <- tolower( tweets.df$username )
  
  wclist <- read.delim("/home/ec2-user/SageMaker/CategoryKeyword", sep = "\t", header = TRUE)
  wclist$Category <- tolower( wclist$Category )
  wclist$Keyword  <- tolower( wclist$Keyword )
  
  keywords <- trimws( wclist[ wclist$Category == tolower(category), "Keyword"] )
  
  
  
  if( length(grep(paste(keywords,collapse="|"), 
           tolower(tweets.df$username)) ) != 0){
    tweets.df <- tweets.df[ -c( (grep(paste(keywords,collapse="|"), 
                                      tweets.df$username))), ]
  }

  return( tweets.df)
  
}


############################################################
# To remove tweets with keywords from multiple cateogories #
############################################################

tweetsUniqueCategory <- function( tweets.df, category){
    
    wclist <- read.delim("/home/ec2-user/SageMaker/CategoryKeyword", sep = "\t", header = TRUE)
    wclist$Category <- tolower( wclist$Category )
    wclist$Keyword  <- tolower( wclist$Keyword )
    
    if( category == "LNG IUD"| category == "Copper IUD"){
          excludedKeywords <- trimws( wclist[ ( wclist$Category != tolower(category) & wclist$Category != tolower("IUD") ), "Keyword"] )
          includedKeywords <- trimws( wclist[ wclist$Category == tolower(category), "Keyword"] )
    }else{
          excludedKeywords <- trimws( wclist[ wclist$Category != tolower(category), "Keyword"] )
          includedKeywords <- trimws( wclist[ wclist$Category == tolower(category), "Keyword"] )
    }
    
  
  excluded.tweets <- grep(paste(excludedKeywords,collapse="|"), 
                                    tweets.df$text)
  included.tweets <- grep(paste(includedKeywords,collapse="|"), 
                          tweets.df$text)
  
  final.tweets <- included.tweets[ ! included.tweets %in% excluded.tweets ]
  
  if( length( final.tweets) != 0 ){
    tweets.df <- tweets.df[ c( final.tweets ) , ]
  }

  return( tweets.df)
  
}


The previous set of functions are applied to each of the 8 birth control categories. Additionally, we add some steps to:
- remove duplicated tweets. 
- remove male contractive tweets based on three expressions: "male contraception", "male contraceptive"and "male birth control"
- remove emergency contraceptive tweets based on three expressions: "emergency contraception", "emergency contraceptive"and "emergency birth control"

##############
# Copper IUD #
##############

#1st. Unify the tweets
tweets <-  unifyFiles( folder_path = "/home/ec2-user/SageMaker/Tweets/", 
                       folder_name = "Copper_IUD/")

tweets <- tweets[ tweets$text != "", ]

folder_name = "Copper_IUD/"
print( paste0( "In total we have ", nrow( tweets ), " initial tweets for the category ", folder_name ))

tweets <- tweets[!duplicated(tweets$text),]

folder_name = "Copper_IUD/"
print( paste0( "Without duplicates, in total we have ", nrow( tweets ), " initial tweets for the category ", folder_name ))

tweets$text <- as.character( tweets$text )

copperNoClean <- tweets

#2nd. remove male BC, then count tweets
maleContraception <- grep(" male contraception", tolower(tweets$text))
maleContraceptive <- grep(" male contraceptive", tolower(tweets$text))
maleBirthControl <- grep(" male birth control", tolower(tweets$text))

maleBCrows <- c(maleContraception, maleContraceptive, maleBirthControl)
print(paste0("There were ", length(maleBCrows), " tweets about male birth control"))

tweets <- tweets[-maleBCrows, ]

folder_name = "Copper_IUD/"
print( paste0( "Without male contraception, in total we have ", nrow( tweets ), " tweets for the category ", folder_name ))

tweets$text <- as.character( tweets$text )

#3rd. remove emergency contraception, then count tweets
emergencyContraception <- grep(" emergency contraception", tolower(tweets$text))
emergencyContraceptive <- grep(" emergency contraceptive", tolower(tweets$text))
emergencyBirthControl <- grep(" emergency birth control", tolower(tweets$text))

emergencyContraceptionrows <- c(emergencyContraception, emergencyContraceptive, emergencyBirthControl)
print(paste0("There were ", length(emergencyContraceptionrows), " tweets about emergency contraception"))

tweets <- tweets[-emergencyContraceptionrows, ]

folder_name = "Copper_IUD/"
print( paste0( "Without emergency contraception, in total we have ", nrow( tweets ), " tweets for the category ", folder_name ))
tweets$text <- as.character( tweets$text )

#4th. Clean tweet text
print("Length of the tweets before cleaning")
summary( nchar( tweets$text ) )
ourCleanTweets <- cleanTweets( tweets.df = tweets )

print("Length of the tweets after cleaning")
summary( nchar( ourCleanTweets$text ) )

#5th. Clean by username
ourCleanTweets <- usernameClean( tweets.df = ourCleanTweets, category = "Copper IUD")
print( paste0( "In total we have ", nrow(ourCleanTweets), " tweets that do not contain the keyword as username" ))

#6th. Remove duplicates
ourCleanTweets <- ourCleanTweets[!duplicated(ourCleanTweets$text),]
print( paste0( "In total we have ", nrow(ourCleanTweets), " tweets without duplicates" ))

#7th. Save the file to make Figures
write.table( ourCleanTweets, 
            file = "/home/ec2-user/SageMaker/SemiCleanAndAggregateTweets/copperIUD_SemiCleanTweets.txt", 
            col.names = TRUE, 
            row.names = FALSE, 
            sep = "\t", 
            quote = FALSE )

copperIUDSemiClean <- ourCleanTweets

#8th. Remove tweets mentioning multiple categories
ourCleanTweets <- tweetsUniqueCategory( tweets.df = ourCleanTweets, category = "Copper IUD")
print( paste0( "In total we have ", nrow(ourCleanTweets), " tweets that refer to a unique category" ))

#9th. Save the file to proceed to the AWS comprehend analysis
write.table( ourCleanTweets, 
            file = "/home/ec2-user/SageMaker/CleanAndAggregateTweets/copperIUD_CleanTweets.txt", 
            col.names = TRUE, 
            row.names = FALSE, 
            sep = "\t", 
            quote = FALSE )

copperIUD <- ourCleanTweets
rm(ourCleanTweets)
rm(tweets)

###########
# LNG IUD #
###########

#1st. Unify the tweets
tweets <-  unifyFiles( folder_path = "/home/ec2-user/SageMaker/Tweets/", 
                       folder_name = "LNG-IUD/")

tweets <- tweets[ tweets$text != "", ]

folder_name = "LNG-IUD/"
print( paste0( "In total we have ", nrow( tweets ), " initial tweets for the category ", folder_name ))

tweets <- tweets[!duplicated(tweets$text),]

folder_name = "LNG-IUD/"
print( paste0( "Without duplicates, in total we have ", nrow( tweets ), " initial tweets for the category ", folder_name ))

tweets$text <- as.character( tweets$text )

lngNoClean <- tweets 

#2nd. remove male BC, then count tweets
maleContraception <- grep(" male contraception", tolower(tweets$text))
maleContraceptive <- grep(" male contraceptive", tolower(tweets$text))
maleBirthControl <- grep(" male birth control", tolower(tweets$text))

maleBCrows <- c(maleContraception, maleContraceptive, maleBirthControl)
print(paste0("There were ", length(maleBCrows), " tweets about male birth control"))

tweets <- tweets[-maleBCrows, ]

folder_name = "LNG-IUD/"
print( paste0( "Without male contraception, in total we have ", nrow( tweets ), " tweets for the category ", folder_name ))

tweets$text <- as.character( tweets$text )

#3rd. remove emergency contraception, then count tweets
emergencyContraception <- grep(" emergency contraception", tolower(tweets$text))
emergencyContraceptive <- grep(" emergency contraceptive", tolower(tweets$text))
emergencyBirthControl <- grep(" emergency birth control", tolower(tweets$text))

emergencyContraceptionrows <- c(emergencyContraception, emergencyContraceptive, emergencyBirthControl)
print(paste0("There were ", length(emergencyContraceptionrows), " tweets about emergency contraception"))

tweets <- tweets[-emergencyContraceptionrows, ]

folder_name = "LNG-IUD/"
print( paste0( "Without emergency contraception, in total we have ", nrow( tweets ), " tweets for the category ", folder_name ))
tweets$text <- as.character( tweets$text )

#4th. Clean tweet text
print("Length of the tweets before cleaning")
summary( nchar( tweets$text ) )
ourCleanTweets <- cleanTweets( tweets.df = tweets )

print("Length of the tweets after cleaning")
summary( nchar( ourCleanTweets$text ) )

#5th. Clean by username
ourCleanTweets <- usernameClean( tweets.df = ourCleanTweets, category = "LNG IUD")
print( paste0( "In total we have ", nrow(ourCleanTweets), " tweets that do not contain the keyword as username" ))

#6th. Remove duplicates
ourCleanTweets <- ourCleanTweets[!duplicated(ourCleanTweets$text),]
print( paste0( "In total we have ", nrow(ourCleanTweets), " tweets without duplicates" ))

#7th. Save the file to make Figures
write.table( ourCleanTweets, 
            file = "/home/ec2-user/SageMaker/SemiCleanAndAggregateTweets/LNG-IUD_SemiCleanTweets.txt", 
            col.names = TRUE, 
            row.names = FALSE, 
            sep = "\t", 
            quote = FALSE )

LNG_IUDSemiClean <- ourCleanTweets

#8th. Remove tweets mentioning multiple categories
ourCleanTweets <- tweetsUniqueCategory( tweets.df = ourCleanTweets, category = "LNG IUD")
print( paste0( "In total we have ", nrow(ourCleanTweets), " tweets that refer to a unique category" ))

#9th. Save the file to proceed to the AWS comprehend analysis
write.table( ourCleanTweets, 
            file = "/home/ec2-user/SageMaker/CleanAndAggregateTweets/LNG-IUD_CleanTweets.txt", 
            col.names = TRUE, 
            row.names = FALSE, 
            sep = "\t", 
            quote = FALSE )

LNG_IUD <- ourCleanTweets
rm(ourCleanTweets)
rm(tweets)

#######
# IUD #
#######

#1st. Unify the tweets
tweets <-  unifyFiles( folder_path = "/home/ec2-user/SageMaker/Tweets/", 
                       folder_name = "IUD/")

tweets <- tweets[ tweets$text != "", ]

folder_name = "IUD/"
print( paste0( "In total we have ", nrow( tweets ), " initial tweets for the category ", folder_name ))

tweets <- tweets[!duplicated(tweets$text),]

folder_name = "IUD/"
print( paste0( "Without duplicates, in total we have ", nrow( tweets ), " initial tweets for the category ", folder_name ))

tweets$text <- as.character( tweets$text )

IUD_NoClean <- tweets

#2nd. remove male BC, then count tweets
maleContraception <- grep(" male contraception", tolower(tweets$text))
maleContraceptive <- grep(" male contraceptive", tolower(tweets$text))
maleBirthControl <- grep(" male birth control", tolower(tweets$text))

maleBCrows <- c(maleContraception, maleContraceptive, maleBirthControl)
print(paste0("There were ", length(maleBCrows), " tweets about male birth control"))

tweets <- tweets[-maleBCrows, ]

folder_name = "IUD/"
print( paste0( "Without male contraception, in total we have ", nrow( tweets ), " tweets for the category ", folder_name ))

tweets$text <- as.character( tweets$text )

#3rd. remove emergency contraception, then count tweets
emergencyContraception <- grep(" emergency contraception", tolower(tweets$text))
emergencyContraceptive <- grep(" emergency contraceptive", tolower(tweets$text))
emergencyBirthControl <- grep(" emergency birth control", tolower(tweets$text))

emergencyContraceptionrows <- c(emergencyContraception, emergencyContraceptive, emergencyBirthControl)
print(paste0("There were ", length(emergencyContraceptionrows), " tweets about emergency contraception"))

tweets <- tweets[-emergencyContraceptionrows, ]

folder_name = "IUD/"
print( paste0( "Without emergency contraception, in total we have ", nrow( tweets ), " tweets for the category ", folder_name ))
tweets$text <- as.character( tweets$text )

#4th. Clean tweet text
print("Length of the tweets before cleaning")
summary( nchar( tweets$text ) )
ourCleanTweets <- cleanTweets( tweets.df = tweets )

print("Length of the tweets after cleaning")
summary( nchar( ourCleanTweets$text ) )

#5th. Clean by username
ourCleanTweets <- usernameClean( tweets.df = ourCleanTweets, category = "IUD")
print( paste0( "In total we have ", nrow(ourCleanTweets), " tweets that do not contain the keyword as username" ))

#6th. Remove duplicates
ourCleanTweets <- ourCleanTweets[!duplicated(ourCleanTweets$text),]
print( paste0( "In total we have ", nrow(ourCleanTweets), " tweets without duplicates" ))

#7th. Save the file to make Figures
write.table( ourCleanTweets, 
            file = "/home/ec2-user/SageMaker/SemiCleanAndAggregateTweets/IUD_SemiCleanTweets.txt", 
            col.names = TRUE, 
            row.names = FALSE, 
            sep = "\t", 
            quote = FALSE )

IUDSemiClean <- ourCleanTweets

#8th. Remove tweets mentioning multiple categories
ourCleanTweets <- tweetsUniqueCategory( tweets.df = ourCleanTweets, category = "IUD")
print( paste0( "In total we have ", nrow(ourCleanTweets), " tweets that refer to a unique category" ))

#9th. Save the file to proceed to the AWS comprehend analysis
write.table( ourCleanTweets, 
            file = "/home/ec2-user/SageMaker/CleanAndAggregateTweets/IUD_CleanTweets.txt", 
            col.names = TRUE, 
            row.names = FALSE, 
            sep = "\t", 
            quote = FALSE )

IUD <- ourCleanTweets
rm(ourCleanTweets)
rm(tweets)

###########
# Implant #
###########

#1st. Unify the tweets
tweets <-  unifyFiles( folder_path = "/home/ec2-user/SageMaker/Tweets/", 
                       folder_name = "Implant/")

tweets <- tweets[ tweets$text != "", ]

folder_name = "Implant/"
print( paste0( "In total we have ", nrow( tweets ), " initial tweets for the category ", folder_name ))

tweets <- tweets[!duplicated(tweets$text),]

folder_name = "Implant/"
print( paste0( "Without duplicates, in total we have ", nrow( tweets ), " initial tweets for the category ", folder_name ))

tweets$text <- as.character( tweets$text )

implantNoClean <- tweets 

#2nd. remove male BC, then count tweets
maleContraception <- grep(" male contraception", tolower(tweets$text))
maleContraceptive <- grep(" male contraceptive", tolower(tweets$text))
maleBirthControl <- grep(" male birth control", tolower(tweets$text))

maleBCrows <- c(maleContraception, maleContraceptive, maleBirthControl)
print(paste0("There were ", length(maleBCrows), " tweets about male birth control"))

tweets <- tweets[-maleBCrows, ]

folder_name = "Implant/"
print( paste0( "Without male contraception, in total we have ", nrow( tweets ), " tweets for the category ", folder_name ))

tweets$text <- as.character( tweets$text )

#3rd. remove emergency contraception, then count tweets
emergencyContraception <- grep(" emergency contraception", tolower(tweets$text))
emergencyContraceptive <- grep(" emergency contraceptive", tolower(tweets$text))
emergencyBirthControl <- grep(" emergency birth control", tolower(tweets$text))

emergencyContraceptionrows <- c(emergencyContraception, emergencyContraceptive, emergencyBirthControl)
print(paste0("There were ", length(emergencyContraceptionrows), " tweets about emergency contraception"))

tweets <- tweets[-emergencyContraceptionrows, ]

folder_name = "Implant/"
print( paste0( "Without emergency contraception, in total we have ", nrow( tweets ), " tweets for the category ", folder_name ))
tweets$text <- as.character( tweets$text )

#4th. Clean tweet text
print("Length of the tweets before cleaning")
summary( nchar( tweets$text ) )
ourCleanTweets <- cleanTweets( tweets.df = tweets )

print("Length of the tweets after cleaning")
summary( nchar( ourCleanTweets$text ) )

#5th. Clean by username
ourCleanTweets <- usernameClean( tweets.df = ourCleanTweets, category = "Implant")
print( paste0( "In total we have ", nrow(ourCleanTweets), " tweets that do not contain the keyword as username" ))

#6th. Remove duplicates
ourCleanTweets <- ourCleanTweets[!duplicated(ourCleanTweets$text),]
print( paste0( "In total we have ", nrow(ourCleanTweets), " tweets without duplicates" ))

#7th. Save the file to make Figures
write.table( ourCleanTweets, 
            file = "/home/ec2-user/SageMaker/SemiCleanAndAggregateTweets/Implant_SemiCleanTweets.txt", 
            col.names = TRUE, 
            row.names = FALSE, 
            sep = "\t", 
            quote = FALSE )

ImplantSemiClean <- ourCleanTweets

#8th. Remove tweets mentioning multiple categories
ourCleanTweets <- tweetsUniqueCategory( tweets.df = ourCleanTweets, category = "Implant")
print( paste0( "In total we have ", nrow(ourCleanTweets), " tweets that refer to a unique category" ))

#9th. Save the file to proceed to the AWS comprehend analysis
write.table( ourCleanTweets, 
            file = "/home/ec2-user/SageMaker/CleanAndAggregateTweets/Implant_CleanTweets.txt", 
            col.names = TRUE, 
            row.names = FALSE, 
            sep = "\t", 
            quote = FALSE )

Implant <- ourCleanTweets
rm(ourCleanTweets)
rm(tweets)

###########
# TheShot #
###########

#1st. Unify the tweets
tweets <-  unifyFiles( folder_path = "/home/ec2-user/SageMaker/Tweets/", 
                       folder_name = "TheShot/")

tweets <- tweets[ tweets$text != "", ]

folder_name = "TheShot/"
print( paste0( "In total we have ", nrow( tweets ), " initial tweets for the category ", folder_name ))

tweets <- tweets[!duplicated(tweets$text),]

folder_name = "TheShot/"
print( paste0( "Without duplicates, in total we have ", nrow( tweets ), " initial tweets for the category ", folder_name ))

tweets$text <- as.character( tweets$text )

shotNoClean <- tweets

#2nd. remove male BC, then count tweets
maleContraception <- grep(" male contraception", tolower(tweets$text))
maleContraceptive <- grep(" male contraceptive", tolower(tweets$text))
maleBirthControl <- grep(" male birth control", tolower(tweets$text))

maleBCrows <- c(maleContraception, maleContraceptive, maleBirthControl)
print(paste0("There were ", length(maleBCrows), " tweets about male birth control"))

tweets <- tweets[-maleBCrows, ]

folder_name = "TheShot/"
print( paste0( "Without male contraception, in total we have ", nrow( tweets ), " tweets for the category ", folder_name ))

tweets$text <- as.character( tweets$text )

#3rd. remove emergency contraception, then count tweets
emergencyContraception <- grep(" emergency contraception", tolower(tweets$text))
emergencyContraceptive <- grep(" emergency contraceptive", tolower(tweets$text))
emergencyBirthControl <- grep(" emergency birth control", tolower(tweets$text))

emergencyContraceptionrows <- c(emergencyContraception, emergencyContraceptive, emergencyBirthControl)
print(paste0("There were ", length(emergencyContraceptionrows), " tweets about emergency contraception"))

tweets <- tweets[-emergencyContraceptionrows, ]

folder_name = "TheShot/"
print( paste0( "Without emergency contraception, in total we have ", nrow( tweets ), " tweets for the category ", folder_name ))
tweets$text <- as.character( tweets$text )

#4th. Clean tweet text
print("Length of the tweets before cleaning")
summary( nchar( tweets$text ) )
ourCleanTweets <- cleanTweets( tweets.df = tweets )

print("Length of the tweets after cleaning")
summary( nchar( ourCleanTweets$text ) )

#5th. Clean by username
ourCleanTweets <- usernameClean( tweets.df = ourCleanTweets, category = "Shot")
print( paste0( "In total we have ", nrow(ourCleanTweets), " tweets that do not contain the keyword as username" ))

#6th. Remove duplicates
ourCleanTweets <- ourCleanTweets[!duplicated(ourCleanTweets$text),]
print( paste0( "In total we have ", nrow(ourCleanTweets), " tweets without duplicates" ))

#7th. Save the file to make Figures
write.table( ourCleanTweets, 
            file = "/home/ec2-user/SageMaker/SemiCleanAndAggregateTweets/Shot_SemiCleanTweets.txt", 
            col.names = TRUE, 
            row.names = FALSE, 
            sep = "\t", 
            quote = FALSE )

shotSemiClean <- ourCleanTweets

#8th. Remove tweets mentioning multiple categories
ourCleanTweets <- tweetsUniqueCategory( tweets.df = ourCleanTweets, category = "Shot")
print( paste0( "In total we have ", nrow(ourCleanTweets), " tweets that refer to a unique category" ))

#9th. Save the file to proceed to the AWS comprehend analysis
write.table( ourCleanTweets, 
            file = "/home/ec2-user/SageMaker/CleanAndAggregateTweets/Shot_CleanTweets.txt", 
            col.names = TRUE, 
            row.names = FALSE, 
            sep = "\t", 
            quote = FALSE )

shot <- ourCleanTweets
rm(ourCleanTweets)
rm(tweets)

############
# The pill #
############

#1st. Unify the tweets
tweets <-  unifyFiles( folder_path = "/home/ec2-user/SageMaker/Tweets/", 
                       folder_name = "ThePill/")

tweets <- tweets[ tweets$text != "", ]

folder_name = "ThePill/"
print( paste0( "In total we have ", nrow( tweets ), " initial tweets for the category ", folder_name ))

tweets <- tweets[!duplicated(tweets$text),]

folder_name = "ThePill/"
print( paste0( "Without duplicates, in total we have ", nrow( tweets ), " initial tweets for the category ", folder_name ))

tweets$text <- as.character( tweets$text )

pillNoClean <- tweets

#2nd. remove male BC, then count tweets
maleContraception <- grep(" male contraception", tolower(tweets$text))
maleContraceptive <- grep(" male contraceptive", tolower(tweets$text))
maleBirthControl <- grep(" male birth control", tolower(tweets$text))

maleBCrows <- c(maleContraception, maleContraceptive, maleBirthControl)
print(paste0("There were ", length(maleBCrows), " tweets about male birth control"))

tweets <- tweets[-maleBCrows, ]

folder_name = "ThePill/"
print( paste0( "Without male contraception, in total we have ", nrow( tweets ), " tweets for the category ", folder_name ))

tweets$text <- as.character( tweets$text )

#3rd. remove emergency contraception, then count tweets
emergencyContraception <- grep(" emergency contraception", tolower(tweets$text))
emergencyContraceptive <- grep(" emergency contraceptive", tolower(tweets$text))
emergencyBirthControl <- grep(" emergency birth control", tolower(tweets$text))

emergencyContraceptionrows <- c(emergencyContraception, emergencyContraceptive, emergencyBirthControl)
print(paste0("There were ", length(emergencyContraceptionrows), " tweets about emergency contraception"))

tweets <- tweets[-emergencyContraceptionrows, ]

folder_name = "ThePill/"
print( paste0( "Without emergency contraception, in total we have ", nrow( tweets ), " tweets for the category ", folder_name ))
tweets$text <- as.character( tweets$text )

#4th. Clean tweet text
print("Length of the tweets before cleaning")
summary( nchar( tweets$text ) )
ourCleanTweets <- cleanTweets( tweets.df = tweets )

print("Length of the tweets after cleaning")
summary( nchar( ourCleanTweets$text ) )

#5th. Clean by username
ourCleanTweets <- usernameClean( tweets.df = ourCleanTweets, category = "Pill")
print( paste0( "In total we have ", nrow(ourCleanTweets), " tweets that do not contain the keyword as username" ))

#6th. Remove duplicates
ourCleanTweets <- ourCleanTweets[!duplicated(ourCleanTweets$text),]
print( paste0( "In total we have ", nrow(ourCleanTweets), " tweets without duplicates" ))

#7th. Save the file to make Figures
write.table( ourCleanTweets, 
            file = "/home/ec2-user/SageMaker/SemiCleanAndAggregateTweets/Pill_SemiCleanTweets.txt", 
            col.names = TRUE, 
            row.names = FALSE, 
            sep = "\t", 
            quote = FALSE )

pillSemiClean <- ourCleanTweets

#8th. Remove tweets mentioning multiple categories
ourCleanTweets <- tweetsUniqueCategory( tweets.df = ourCleanTweets, category = "Pill")
print( paste0( "In total we have ", nrow(ourCleanTweets), " tweets that refer to a unique category" ))

#9th. Save the file to proceed to the AWS comprehend analysis
write.table( ourCleanTweets, 
            file = "/home/ec2-user/SageMaker/CleanAndAggregateTweets/Pill_CleanTweets.txt", 
            col.names = TRUE, 
            row.names = FALSE, 
            sep = "\t", 
            quote = FALSE )

pill <- ourCleanTweets
rm(ourCleanTweets)
rm(tweets)

############
# The Ring #
############

#1st. Unify the tweets
tweets <-  unifyFiles( folder_path = "/home/ec2-user/SageMaker/Tweets/", 
                       folder_name = "TheRing/")

tweets <- tweets[ tweets$text != "", ]

folder_name = "TheRing/"
print( paste0( "In total we have ", nrow( tweets ), " initial tweets for the category ", folder_name ))

tweets <- tweets[!duplicated(tweets$text),]

folder_name = "TheRing/"
print( paste0( "Without duplicates, in total we have ", nrow( tweets ), " initial tweets for the category ", folder_name ))

tweets$text <- as.character( tweets$text )

ringNoClean <- tweets

#2nd. remove male BC, then count tweets
maleContraception <- grep(" male contraception", tolower(tweets$text))
maleContraceptive <- grep(" male contraceptive", tolower(tweets$text))
maleBirthControl <- grep(" male birth control", tolower(tweets$text))

maleBCrows <- c(maleContraception, maleContraceptive, maleBirthControl)
print(paste0("There were ", length(maleBCrows), " tweets about male birth control"))

tweets <- tweets[-maleBCrows, ]

folder_name = "TheRing/"
print( paste0( "Without male contraception, in total we have ", nrow( tweets ), " tweets for the category ", folder_name ))

tweets$text <- as.character( tweets$text )

#3rd. remove emergency contraception, then count tweets
emergencyContraception <- grep(" emergency contraception", tolower(tweets$text))
emergencyContraceptive <- grep(" emergency contraceptive", tolower(tweets$text))
emergencyBirthControl <- grep(" emergency birth control", tolower(tweets$text))

emergencyContraceptionrows <- c(emergencyContraception, emergencyContraceptive, emergencyBirthControl)
print(paste0("There were ", length(emergencyContraceptionrows), " tweets about emergency contraception"))

tweets <- tweets[-emergencyContraceptionrows, ]

folder_name = "TheRing/"
print( paste0( "Without emergency contraception, in total we have ", nrow( tweets ), " tweets for the category ", folder_name ))
tweets$text <- as.character( tweets$text )

#4th. Clean tweet text
print("Length of the tweets before cleaning")
summary( nchar( tweets$text ) )
ourCleanTweets <- cleanTweets( tweets.df = tweets )

print("Length of the tweets after cleaning")
summary( nchar( ourCleanTweets$text ) )

#5th. Clean by username
ourCleanTweets <- usernameClean( tweets.df = ourCleanTweets, category = "Ring")
print( paste0( "In total we have ", nrow(ourCleanTweets), " tweets that do not contain the keyword as username" ))

#6th. Remove duplicates
ourCleanTweets <- ourCleanTweets[!duplicated(ourCleanTweets$text),]
print( paste0( "In total we have ", nrow(ourCleanTweets), " tweets without duplicates" ))

#7th. Save the file to make Figures
write.table( ourCleanTweets, 
            file = "/home/ec2-user/SageMaker/SemiCleanAndAggregateTweets/Ring_SemiCleanTweets.txt", 
            col.names = TRUE, 
            row.names = FALSE, 
            sep = "\t", 
            quote = FALSE )

ringSemiClean <- ourCleanTweets

#8th. Remove tweets mentioning multiple categories
ourCleanTweets <- tweetsUniqueCategory( tweets.df = ourCleanTweets, category = "Ring")
print( paste0( "In total we have ", nrow(ourCleanTweets), " tweets that refer to a unique category" ))

#9th. Save the file to proceed to the AWS comprehend analysis
write.table( ourCleanTweets, 
            file = "/home/ec2-user/SageMaker/CleanAndAggregateTweets/Ring_CleanTweets.txt", 
            col.names = TRUE, 
            row.names = FALSE, 
            sep = "\t", 
            quote = FALSE )

ring <- ourCleanTweets
rm(ourCleanTweets)
rm(tweets)

#############
# The Patch #
############

#1st. Unify the tweets
tweets <-  unifyFiles( folder_path = "/home/ec2-user/SageMaker/Tweets/", 
                       folder_name = "ThePatch/")

tweets <- tweets[ tweets$text != "", ]

folder_name = "ThePatch/"
print( paste0( "In total we have ", nrow( tweets ), " initial tweets for the category ", folder_name ))

tweets <- tweets[!duplicated(tweets$text),]

folder_name = "ThePatch/"
print( paste0( "Without duplicates, in total we have ", nrow( tweets ), " initial tweets for the category ", folder_name ))

tweets$text <- as.character( tweets$text )

patchNoClean <- tweets

#2nd. remove male BC, then count tweets
maleContraception <- grep(" male contraception", tolower(tweets$text))
maleContraceptive <- grep(" male contraceptive", tolower(tweets$text))
maleBirthControl <- grep(" male birth control", tolower(tweets$text))

maleBCrows <- c(maleContraception, maleContraceptive, maleBirthControl)
print(paste0("There were ", length(maleBCrows), " tweets about male birth control"))

tweets <- tweets[-maleBCrows, ]

folder_name = "ThePatch/"
print( paste0( "Without male contraception, in total we have ", nrow( tweets ), " tweets for the category ", folder_name ))

tweets$text <- as.character( tweets$text )

#3rd. remove emergency contraception, then count tweets
emergencyContraception <- grep(" emergency contraception", tolower(tweets$text))
emergencyContraceptive <- grep(" emergency contraceptive", tolower(tweets$text))
emergencyBirthControl <- grep(" emergency birth control", tolower(tweets$text))

emergencyContraceptionrows <- c(emergencyContraception, emergencyContraceptive, emergencyBirthControl)
print(paste0("There were ", length(emergencyContraceptionrows), " tweets about emergency contraception"))

tweets <- tweets[-emergencyContraceptionrows, ]

folder_name = "ThePatch/"
print( paste0( "Without emergency contraception, in total we have ", nrow( tweets ), " tweets for the category ", folder_name ))
tweets$text <- as.character( tweets$text )

#4th. Clean tweet text
print("Length of the tweets before cleaning")
summary( nchar( tweets$text ) )
ourCleanTweets <- cleanTweets( tweets.df = tweets )

print("Length of the tweets after cleaning")
summary( nchar( ourCleanTweets$text ) )

#5th. Clean by username
ourCleanTweets <- usernameClean( tweets.df = ourCleanTweets, category = "Patch")
print( paste0( "In total we have ", nrow(ourCleanTweets), " tweets that do not contain the keyword as username" ))

#6th. Remove duplicates
ourCleanTweets <- ourCleanTweets[!duplicated(ourCleanTweets$text),]
print( paste0( "In total we have ", nrow(ourCleanTweets), " tweets without duplicates" ))

#7th. Save the file to make Figures
write.table( ourCleanTweets, 
            file = "/home/ec2-user/SageMaker/SemiCleanAndAggregateTweets/Patch_SemiCleanTweets.txt", 
            col.names = TRUE, 
            row.names = FALSE, 
            sep = "\t", 
            quote = FALSE )

patchSemiClean <- ourCleanTweets

#8th. Remove tweets mentioning multiple categories
ourCleanTweets <- tweetsUniqueCategory( tweets.df = ourCleanTweets, category = "Patch")
print( paste0( "In total we have ", nrow(ourCleanTweets), " tweets that refer to a unique category" ))

#9th. Save the file to proceed to the AWS comprehend analysis
write.table( ourCleanTweets, 
            file = "/home/ec2-user/SageMaker/CleanAndAggregateTweets/Patch_CleanTweets.txt", 
            col.names = TRUE, 
            row.names = FALSE, 
            sep = "\t", 
            quote = FALSE )

patch <- ourCleanTweets
rm(ourCleanTweets)

#####################################
## Summary of the tweets extracted ##
#####################################

#all tweets extracted
totalTweets <- rbind( copperNoClean, lngNoClean, IUD_NoClean,
                     implantNoClean,shotNoClean,pillNoClean,
                     ringNoClean, patchNoClean)



dim(totalTweets)

sessionInfo()
