==> 3.sentimentAnalysisAmazonComprehend.ipynb <==
# Sentiment analysis using Amazon Comprehend

This notebook contains the script to perform a sentiment analysis using Amazon Comprehend. 
We will run the sentiment analysis for all the clean tweets (tweets with only keywords of one category) to extract the sentiment towards that BC category. 

When running this notebook, we saved the output in "AWScomprenhend_output"

import pandas as pd
from collections import OrderedDict
import requests
import boto3 
comprehend = boto3.client('comprehend', region_name='us-east-1')
import os
path=os.getcwd()
import timeit

**Note that the average time to process one tweet is 0.0591 seconds**

###########
# LNG IUD #
###########

df = pd.read_csv('/home/ec2-user/SageMaker/CleanAndAggregateTweets/LNG-IUD_CleanTweets.txt',
                 sep = '\t' )
len(df.text)

start = timeit.default_timer()
dfTweet = pd.DataFrame(columns=["tweets" ,"sentiments" ,"positive" ,"negative" ,"neutral", "mixed" ])


for i in range(len(df.text)):  
    #print(i) 
    if pd.notna(df.text[i]):
        res = comprehend.detect_sentiment(Text=df.text[i] , LanguageCode='en')
        s = res.get('Sentiment')
        p = res.get('SentimentScore')['Positive']
        neg = res.get('SentimentScore')['Negative']
        neu = res.get('SentimentScore')['Neutral']
        mix = res.get('SentimentScore')['Mixed']

        
        dfTweet = dfTweet.append({"tweets": df.text[i],"sentiments": s, 'positive': p, 'negative': neg, 
                                 'neutral': neu, 'mixed': mix},ignore_index=True)

stop = timeit.default_timer()

print('Time: ', stop - start)  

dfTweet.to_csv(path_or_buf=path+'/'+"AWScomprenhend_output/LNG-IUD_AWScomprehend_complete.csv")

##############
# Copper IUD #
##############

df = pd.read_csv('/home/ec2-user/SageMaker/CleanAndAggregateTweets/copperIUD_CleanTweets.txt',
                 sep = '\t' )
len(df.text)

start = timeit.default_timer()
dfTweet = pd.DataFrame(columns=["tweets" ,"sentiments" ,"positive" ,"negative" ,"neutral", "mixed" ])


for i in range(len(df.text)):  
    #print(i) 
    if pd.notna(df.text[i]):
        res = comprehend.detect_sentiment(Text=df.text[i] , LanguageCode='en')
        s = res.get('Sentiment')
        p = res.get('SentimentScore')['Positive']
        neg = res.get('SentimentScore')['Negative']
        neu = res.get('SentimentScore')['Neutral']
        mix = res.get('SentimentScore')['Mixed']

        
        dfTweet = dfTweet.append({"tweets": df.text[i],"sentiments": s, 'positive': p, 'negative': neg, 
                                 'neutral': neu, 'mixed': mix},ignore_index=True)

stop = timeit.default_timer()


print('Time: ', stop - start)  

dfTweet.to_csv(path_or_buf=path+'/'+"AWScomprenhend_output/copperIUD_AWScomprehend_complete.csv")

#############
# The Patch #
############
df = pd.read_csv('/home/ec2-user/SageMaker/CleanAndAggregateTweets/Patch_CleanTweets.txt',
                 sep = '\t' )
len(df.text)

start = timeit.default_timer()
dfTweet = pd.DataFrame(columns=["tweets" ,"sentiments" ,"positive" ,"negative" ,"neutral", "mixed" ])


for i in range(len(df.text)):  
    #print(i) 
    if pd.notna(df.text[i]):
        res = comprehend.detect_sentiment(Text=df.text[i] , LanguageCode='en')
        s = res.get('Sentiment')
        p = res.get('SentimentScore')['Positive']
        neg = res.get('SentimentScore')['Negative']
        neu = res.get('SentimentScore')['Neutral']
        mix = res.get('SentimentScore')['Mixed']

        
        dfTweet = dfTweet.append({"tweets": df.text[i],"sentiments": s, 'positive': p, 'negative': neg, 
                                 'neutral': neu, 'mixed': mix},ignore_index=True)

stop = timeit.default_timer()


print('Time: ', stop - start)  

dfTweet.to_csv(path_or_buf=path+'/'+"AWScomprenhend_output/Patch_AWScomprehend_complete.csv")

############
# The Ring #
############
df = pd.read_csv('/home/ec2-user/SageMaker/CleanAndAggregateTweets/Ring_CleanTweets.txt',
                 sep = '\t' )
len(df.text)

start = timeit.default_timer()
dfTweet = pd.DataFrame(columns=["tweets" ,"sentiments" ,"positive" ,"negative" ,"neutral", "mixed" ])


for i in range(len(df.text)):  
    #print(i) 
    if pd.notna(df.text[i]):
        res = comprehend.detect_sentiment(Text=df.text[i] , LanguageCode='en')
        s = res.get('Sentiment')
        p = res.get('SentimentScore')['Positive']
        neg = res.get('SentimentScore')['Negative']
        neu = res.get('SentimentScore')['Neutral']
        mix = res.get('SentimentScore')['Mixed']

        
        dfTweet = dfTweet.append({"tweets": df.text[i],"sentiments": s, 'positive': p, 'negative': neg, 
                                 'neutral': neu, 'mixed': mix},ignore_index=True)

stop = timeit.default_timer()

print('Time: ', stop - start)  

dfTweet.to_csv(path_or_buf=path+'/'+"AWScomprenhend_output/Ring_AWScomprehend_complete.csv")

###########
# Implant #
###########
df = pd.read_csv('/home/ec2-user/SageMaker/CleanAndAggregateTweets/Implant_CleanTweets.txt',
                 sep = '\t' )
len(df.text)

start = timeit.default_timer()
dfTweet = pd.DataFrame(columns=["tweets" ,"sentiments" ,"positive" ,"negative" ,"neutral", "mixed" ])


for i in range(len(df.text)):  
    #print(i) 
    if pd.notna(df.text[i]):
        res = comprehend.detect_sentiment(Text=df.text[i] , LanguageCode='en')
        s = res.get('Sentiment')
        p = res.get('SentimentScore')['Positive']
        neg = res.get('SentimentScore')['Negative']
        neu = res.get('SentimentScore')['Neutral']
        mix = res.get('SentimentScore')['Mixed']

        
        dfTweet = dfTweet.append({"tweets": df.text[i],"sentiments": s, 'positive': p, 'negative': neg, 
                                 'neutral': neu, 'mixed': mix},ignore_index=True)

stop = timeit.default_timer()


print('Time: ', stop - start)  

dfTweet.to_csv(path_or_buf=path+'/'+"AWScomprenhend_output/Implant_AWScomprehend_complete.csv")

############
# The pill #
############

df = pd.read_csv('/home/ec2-user/SageMaker/CleanAndAggregateTweets/Pill_CleanTweets.txt',
                 sep = '\t' )
len(df.text)

start = timeit.default_timer()
dfTweet = pd.DataFrame(columns=["tweets" ,"sentiments" ,"positive" ,"negative" ,"neutral", "mixed" ])


for i in range(len(df.text)):  
    #print(i) 
    if pd.notna(df.text[i]):
        res = comprehend.detect_sentiment(Text=df.text[i] , LanguageCode='en')
        s = res.get('Sentiment')
        p = res.get('SentimentScore')['Positive']
        neg = res.get('SentimentScore')['Negative']
        neu = res.get('SentimentScore')['Neutral']
        mix = res.get('SentimentScore')['Mixed']

        
        dfTweet = dfTweet.append({"tweets": df.text[i],"sentiments": s, 'positive': p, 'negative': neg, 
                                 'neutral': neu, 'mixed': mix},ignore_index=True)

stop = timeit.default_timer()


print('Time: ', stop - start)  

dfTweet.to_csv(path_or_buf=path+'/'+"AWScomprenhend_output/Pill_AWScomprehend_complete.csv")

###########
# TheShot #
###########
df = pd.read_csv('/home/ec2-user/SageMaker/CleanAndAggregateTweets/Shot_CleanTweets.txt',
                 sep = '\t' )
len(df.text)

start = timeit.default_timer()
dfTweet = pd.DataFrame(columns=["tweets" ,"sentiments" ,"positive" ,"negative" ,"neutral", "mixed" ])


for i in range(len(df.text)):  
    #print(i) 
    if pd.notna(df.text[i]):
        res = comprehend.detect_sentiment(Text=df.text[i] , LanguageCode='en')
        s = res.get('Sentiment')
        p = res.get('SentimentScore')['Positive']
        neg = res.get('SentimentScore')['Negative']
        neu = res.get('SentimentScore')['Neutral']
        mix = res.get('SentimentScore')['Mixed']

        
        dfTweet = dfTweet.append({"tweets": df.text[i],"sentiments": s, 'positive': p, 'negative': neg, 
                                 'neutral': neu, 'mixed': mix},ignore_index=True)

stop = timeit.default_timer()

print('Time: ', stop - start)  

dfTweet.to_csv(path_or_buf=path+'/'+"AWScomprenhend_output/Shot_AWScomprehend_complete.csv")


#######
# IUD #
#######

df = pd.read_csv('/home/ec2-user/SageMaker/CleanAndAggregateTweets/IUD_CleanTweets.txt',
                 sep = '\t' )
len(df.text)

start = timeit.default_timer()
dfTweet = pd.DataFrame(columns=["tweets" ,"sentiments" ,"positive" ,"negative" ,"neutral", "mixed" ])


for i in range(len(df.text)):  
    #print(i) 
    if pd.notna(df.text[i]):
        res = comprehend.detect_sentiment(Text=df.text[i] , LanguageCode='en')
        s = res.get('Sentiment')
        p = res.get('SentimentScore')['Positive']
        neg = res.get('SentimentScore')['Negative']
        neu = res.get('SentimentScore')['Neutral']
        mix = res.get('SentimentScore')['Mixed']

        
        dfTweet = dfTweet.append({"tweets": df.text[i],"sentiments": s, 'positive': p, 'negative': neg, 
                                 'neutral': neu, 'mixed': mix},ignore_index=True)

stop = timeit.default_timer()


print('Time: ', stop - start)  

dfTweet.to_csv(path_or_buf=path+'/'+"AWScomprenhend_output/IUD_AWScomprehend_complete.csv")
