==> 7.RandomSelection.ipynb <==
# Random selection of tweets

This notebook contains the script and functions to generate random tweets selection according to the number of reviewers for each set and generate the output files in the most convinient format, in this case:
- 5000 tweets per set
- 5 reviewers per group

#####################
# Load Libraries    #
#####################
#install.packages("dplyr")
library("dplyr")

###################################
# Random tweets polarity selection #
###################################
#1st. Look for all the files that are in the folder 
#those tweets are clean tweets that we are using as input for the AWS comprehend polarity analysis
rm(list=ls())
fileList <- list.files("/home/ec2-user/SageMaker/CleanAndAggregateTweets/")

#2nd.we read all the files and we create a unique file with all of them together
for( i in 1:length( fileList ) ){
    if(i == 1){
        mytweetlist <- read.delim(paste0("/home/ec2-user/SageMaker/CleanAndAggregateTweets/", 
                                        fileList[i]), header = TRUE)
    }else{
        newtweetfile <- read.delim(paste0("/home/ec2-user/SageMaker/CleanAndAggregateTweets/", 
                                        fileList[i]), header = TRUE)
        mytweetlist <- rbind( mytweetlist, newtweetfile )
        rm( newtweetfile)
    }
}
print(paste0("In total we have ", nrow(mytweetlist), " tweets for the polarity analysis."))

########################################
# Function to extract tweets randomly #
#######################################
randomTweetsPolarityExtraction <- function( tweets.df, tweetNumber, reviewersName, tweetSet, outputPath ){
    
    randomSelection <- sample_n( tweets.df, tweetNumber+1)
    
    write.table(randomSelection, file= paste0(outputPath, "polarity_set", tweetSet, "WithResults.csv"), col.names = TRUE, row.names = FALSE, sep="," )
    
    randomSelectionForReviewers <- as.data.frame(as.character(randomSelection$text))
    randomSelectionForReviewers$positive <- "-"
    randomSelectionForReviewers$negative <- "-"
    randomSelectionForReviewers$neutral <- "-"
    randomSelectionForReviewers$mixed <- "-"
    randomSelectionForReviewers$falsePositive <- "-"
    
    colnames(randomSelectionForReviewers)[1] <- "text"
    
    for( i in 1:length(reviewersName)){
          write.table(randomSelectionForReviewers, file= paste0(outputPath, reviewersName[i], "_polarity_set", tweetSet,".csv"), col.names = TRUE, row.names = FALSE, sep="," )
    }
}

We have 10 reviewers, so we will create two sets of 500 tweets each. 

randomTweetsPolarityExtraction( tweets.df   = mytweetlist,
                        tweetNumber = 500,
                        reviewersName = c("tomi", "annie", "sonya", "kim", "sofia"),
                        tweetSet = 1,
                        outputPath  = "/home/ec2-user/SageMaker/tweetsToReviewPolarity/" )

randomTweetsPolarityExtraction( tweets.df   = mytweetlist,
                        tweetNumber = 500,
                        reviewersName = c("melody", "anabel", "natalie", "sun", "kathleen", "chloe"),
                        tweetSet = 2,
                        outputPath  = "/home/ec2-user/SageMaker/tweetsToReviewPolarity/" )

sessionInfo()
